FIX Package checks for torchao, EETQ (#2320)
Torchao

Under some unknown circumstances, it can happen that even though
importlib.util.find_spec("torchao") is not None,
importlib_metadata.version("torchao") still fails. This error is now
caught. This error was noticed in the diffusers CI.

EETQ

This is basically a revert of #2226. That PR had to add a check to the
EETQ import as EETQ was broken for some time with latest
transformers (see https://github.com/NetEase-FuXi/EETQ/issues/34 for
context) but that has been fixed.