FIX Revert optimization for LoRA scaling == 1 (#2416)
The PR #2404 introduced an optimization for LoRA in case that scaling ==
1 (see
https://github.com/huggingface/peft/pull/2404#discussion_r1975145200).
This unfortunately leads to recompilation when the model is compiled, as
witnessed by the failing CI here:

https://github.com/huggingface/peft/actions/runs/13755365121/job/38461837691#step:6:157

For now, let's revert the optimization. If we have concrete numbers that
show that the optimization makes a significant difference, we can
start thinking about how to optimize this code path in a
compile-friendly way.