Fix #2422: Modules to save with multiple adapters (#2430)
Using multiple adapters with different `modules_to_save` values leads to a scenario where
it is implicitly assumed that each `ModulesToSaveWrapper` has a module for every loaded adapter.
Since the adapters have different `modules_to_save` values this is not the case and retrieving
the state dict fails with a key lookup error.

In addition to that, after disabling a `ModulesToSaveWrapper`, setting the adapter as active does not
re-enable said adapter.

---------

Co-authored-by: Saeid Ghafouri <s.ghafouri@qub.ac.uk>
Co-authored-by: Benjamin Bossan <BenjaminBossan@users.noreply.github.com>