TST: Refactor remaining common tests to use pytest (#2491)
* Refactor test_adaption_prompt.py

- Did not really use PeftCommonTester, thus removed it
- Removed skip if llama or mistral not avaiable
- Parametrized tests instead of duplicating
- Use small models from Hub instead of creating new ones
- Test coverage misses 3 more lines around loading checkpoint, most
  likely unrelated to adaption prompt but instead due to using hub models
  instead of creating new ones

* Refactor test_feature_extraction.py

Pretty straightforward, test coverage is 100% identical.

* Refactor test_multitask_prompt_tuning

Same arguments apply as for test_adaption_prompt.py

* Refactor test_stablediffusion.py

This was pretty straightforward. After refactoring, the test coverage
was 100% the same.

I noticed, however, that these tests did not cover LoKr, they only
pretended to:

https://github.com/huggingface/peft/blob/37f8dc3458fefb0c1b4362da6733fcf742a1baa3/tests/test_stablediffusion.py#L113-L114

Thus I added LoKr to the test matrix, after which the test coverage if
of course different, but is fine.

* Skip LoKr merging tests when not CUDA

For some reason, the outputs differ after merging. However, I locally
verified that this is already true before this refactor, so let's just
skip for now, as it is out of scope.