FIX Failing target_parameters param usage count (#2676)
For testing target_parameters, we use a tiny Llama4 model. This model
was refactored in
https://github.com/huggingface/transformers/pull/39501, resulting in one
parameter being accessed an additional time:

https://github.com/huggingface/transformers/pull/39501/files#diff-e668ec07f78afdb2cb805d939e47453757f0b9437436cb860fcb7cb2431c9cf5R69

Therefore, a unit test that relied on how often this parameter was
accessed started failing. This PR updates the count to the correct
number.

Additionally debug print statements that were accidentally left over are
now removed.